+++ NEWS.md
@@ -17,6 +17,8 @@
  
  - Fix bug in fixed-effects estimations with multiple LHS and different number of observations per estimation that prevented to get the default behavior for standard-errors to work.
  
+ - Fix bug when using `split` and fixed-effects.
+ 
 # Sun and Abraham (forthcoming, Journal of Econometrics) method for staggered DiD
 
  - For staggered difference-in-difference analyzes: the method of Sun and Abraham (forthcoming, Journal of Econometrics) has been implemented. 
@@ -36,6 +38,8 @@
  - In `etable`, the argument `digits` can now accepts a character value specifying the way the decimals should be displayed. For example if `digits = "r2"` this means that all numbers will be rounded at two decimals and these two decimals will always be displayed. The default behavior is to display significant digits. Follows feature request [#82](https://github.com/lrberge/fixest/issues/82).
  
  - `etable` also gains the argument `digits.stats` which monitors how the fit statistics decimals should be displayed.
+ 
+ - Argument `split` now accepts variable names.
  
 
 # Other